refactor: extract shared Utils::ShellCompletion module
Move completion_shell_parameter and default_completion_shells into a
shared Utils::ShellCompletion module used by both Formula and
Cask::Artifact::GeneratedCompletion, eliminating the duplication.
